fix(payroll): polish scoped payroll access
