+++ apps/webapp/src/app/[locale]/(app)/settings/payroll-access/actions.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from "vitest";
 import { AuthorizationError, ValidationError } from "@/lib/effect/errors";
-import { assertPayrollAccessAdminContext, buildValidatedPayrollAccessInput } from "./actions";
+import { assertPayrollAccessAdminContext, buildValidatedPayrollAccessInput } from "./action-helpers";
 
 vi.mock("next/cache", () => ({
 	revalidatePath: vi.fn(),
